# server/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "基于 Go Fiber + GORM + JWT 的后台管理系统 API 文档",
        "title": "Admin 后台管理系统 API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "admin@example.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/admin/menu/all": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据当前用户角色返回可访问的菜单树(用于前端动态路由)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单"
                ],
                "summary": "获取当前用户的菜单树",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    }
                }
            }
        },
...
        "/admin/system/menu": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "创建新的菜单/目录/按钮",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统管理 - 菜单"
                ],
                "summary": "创建菜单",
                "parameters": [
                    {
                        "description": "菜单信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin.MenuRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.IDResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    }
                }
            }
        },
        "/admin/system/menu/list": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "返回所有菜单的树形结构(用于后台菜单管理页面)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统管理 - 菜单"
                ],
                "summary": "获取菜单管理列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    }
                }
            }
        },
        "/admin/system/menu/name-exists": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "编辑模式下传 id 排除当前记录自身，避免编辑时校验把自己也当作冲突。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统管理 - 菜单"
                ],
                "summary": "检查菜单名称是否存在",
                "parameters": [
                    {
                        "type": "string",
                        "description": "菜单名称",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "当前编辑的菜单ID（编辑模式必传，新增时省略）",
                        "name": "id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/admin/system/menu/path-exists": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "编辑模式下传 id 排除当前记录自身，避免编辑时校验把自己也当作冲突。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统管理 - 菜单"
                ],
                "summary": "检查菜单路径是否存在",
                "parameters": [
                    {
                        "type": "string",
                        "description": "路由路径",
                        "name": "path",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "当前编辑的菜单ID（编辑模式必传，新增时省略）",
                        "name": "id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/admin/system/menu/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "更新菜单信息；父级不能设为自身或其下级",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统管理 - 菜单"
                ],
                "summary": "更新菜单",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "菜单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "菜单信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin.MenuRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "删除指定菜单及其全部下级（含按钮），并清理角色关联，权限码即时失效",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统管理 - 菜单"
                ],
                "summary": "删除菜单",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "菜单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "admin.MenuRequest": {
            "type": "object",
            "required": [
                "name",
                "title",
                "type"
            ],
            "properties": {
                "activeIcon": {
                    "description": "Vben Admin 路由 meta 装饰字段",
                    "type": "string",
                    "example": ""
                },
                "activePath": {
                    "type": "string",
                    "example": ""
                },
                "affixTab": {
                    "type": "boolean"
                },
                "authCode": {
                    "type": "string",
                    "example": "System:Menu:List"
                },
                "badge": {
                    "type": "string",
                    "example": ""
                },
                "badgeType": {
                    "type": "string",
                    "enum": [
                        "dot",
                        "normal"
                    ],
                    "example": ""
                },
                "badgeVariants": {
                    "type": "string",
                    "example": ""
                },
                "component": {
                    "type": "string",
                    "example": "/dashboard/index"
                },
                "hideChildrenInMenu": {
                    "type": "boolean"
                },
                "hideInBreadcrumb": {
                    "type": "boolean"
                },
                "hideInMenu": {
                    "type": "boolean"
                },
                "hideInTab": {
                    "type": "boolean"
                },
                "icon": {
                    "type": "string",
                    "example": "lucide:layout-dashboard"
                },
                "iframeSrc": {
                    "type": "string"
                },
                "keepAlive": {
                    "type": "boolean"
                },
                "link": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "example": "Dashboard"
                },
                "order": {
                    "description": "OrderNo 用指针：区分\"未提交\"和\"显式提交 0\"。\n未提交时 Update 不会重置已有排序值，避免前端少填表单字段就把顺序打乱。",
                    "type": "integer",
                    "example": 1
                },
                "path": {
                    "type": "string",
                    "example": "/dashboard"
                },
                "pid": {
                    "type": "integer",
                    "example": 0
                },
                "redirect": {
                    "type": "string"
                },
                "status": {
                    "type": "integer",
                    "enum": [
                        0,
                        1
                    ],
                    "example": 1
                },
                "title": {
                    "type": "string",
                    "example": "仪表盘"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "catalog",
                        "menu",
                        "button",
                        "embedded",
                        "link"
                    ],
                    "example": "menu"
                }
            }
        },
...
        "dto.IDResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "dto.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {},
                "error": {},
                "message": {
                    "type": "string",
                    "example": "ok"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "输入 Bearer {token} 格式的 JWT 令牌",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}